edit genre

+++ Django/tikitaka/accounts/views.py
@@ -162,15 +162,17 @@ def recommend_bookmark(request, user_id):
 @permission_classes([])
 def edit_profile(request):
     user = User.objects.get(id=request.data.get('id'))
-    user.profile_img = request.data.get('profile_img')
-    user.description = request.data.get('description')
-    genre_id_list = request.data.get('genre_id_list')
-    user.favorite_genres.all().delete()
+    if request.data.get('profile_img'):
+        user.profile_img = request.data.get('profile_img')
+    if request.data.get('description'):
+        user.description = request.data.get('description')
+    genre_id_list = request.POST.getlist('genre_id_list')
     print('@@@@@@@@@@@@@@@@@@@@', genre_id_list)
-    # genre.favorite_users.filter(user=user).delete()
-    for g in genre_id_list:
-        genre = Genre.objects.get(pk=g)
-        genre.favorite_users.add(user)
+    if genre_id_list :
+        user.favorite_genres.all().delete()
+        for g in genre_id_list:
+            genre = Genre.objects.get(pk=g)
+            genre.favorite_users.add(user)
     user.save()
     return Response(status=status.HTTP_200_OK)
